backend/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
astunparse==1.6.3
Async-OAuthlib==0.0.9
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
braceexpand==0.1.7
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
frozenlist==1.3.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
import-expression==1.1.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jishkucord==2.5.2
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.2
priority==2.0.0
py-cord==2.2.2
PyJWT==2.6.0
python-dotenv==0.21.0
quart==0.18.3
Quart-CORS==0.5.0
requests==2.28.1
six==1.16.0
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.1
zipp==3.10.0
